Improve family name search.
The previous family search did a full table scan on the family.name field.

So build an inverted index of families partners names to families. Then
the search can find the lists of families that match the words in the
search query, and from those find the families which match the most words
in the query and return that.