Switch to storing dates as raw strings.
The input GEDCOM file I got from Ruth Pearce has many dates
which are fuzzy; dates with "circa 1946", "June 7 or 8", and
dozens of other variants. There's no way to know the exact
dates, so for ease of import and rendering the data losslessly,
just switch to storing the date as free-text strings. We lose
enforcement of precision there, but the source data isn't
precise anyway.